+++ packages/terra-framework-docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ## 1.20.0 - (April 27, 2023)
 
 * Changed
+  * Updated examples for `terra-popup`.
   * Updated `notification-dialog` and `tabs` examples.
   * Updated examples for `terra-slide-panel-manager`.
   * Updated menu button to use down caret for visual context