+++ packages/terra-popup/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+
+* Changed
+  * Updated close handler to return focus to target object upon pop-up closure. 
+  
 ## 6.65.0 - (April 27, 2023)
 
 * Changed